create application errors

+++ internal/expenses/dto/dto.go
@@ -1,9 +1,9 @@
 package dto
 
 type ExpenseDTO struct {
-	ID          string `json:"id,omitempty"`
-	Amount      int64  `json:"amount"`
-	Description string `json:"description"`
-	Date        string `json:"date"`
-	ExpenseType string `json:"expense_type"`
+	ID          string  `json:"id,omitempty"`
+	Amount      float64 `json:"amount"`
+	Description string  `json:"description"`
+	Date        string  `json:"date"`
+	ExpenseType string  `json:"expense_type"`
 }
